pypolyomino-master/requirements.txt
python-constraint
shapely
svgwrite
